ERM-3711
* create useAgreementsDisplaySettings hook
* use new hook, remove useAgreementsSettings hook
