+++ src/components/IfAccordionIsVisible/IfAccordionIsVisible.js
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types';
 
 import { hiddenAccordions } from '../../constants';
-import { useAgreementsSettings } from '../../hooks';
+import { useAgreementsDisplaySettings } from '../../hooks';
 
 const hiddenAccordionsNames = Object.keys(hiddenAccordions);
 
 const IfAccordionIsVisible = ({ children, name }) => {
-  const settings = useAgreementsSettings();
+  const settings = useAgreementsDisplaySettings();
 
   const isHidden = settings?.parsedSettings.hideAccordions?.[name] || false;
   const isEnabled = !isHidden; // This takes advantage of truthiness to also make sure the path above exists
