+++ src/components/utilities/index.js
@@ -9,6 +9,6 @@ export { default as urls } from './urls';
 export { default as validators } from './validators';
 export { default as parseMclPageSize } from './parseMclPageSize';
 export { default as getTranslatedAcqMethod } from './getTranslatedAcqMethod';
-export { default as parseAgreementSettings } from './parseAgreementSettings';
+export { default as parseAgreementDisplaySettings } from './parseAgreementDisplaySettings';
 export { getRefdataValuesByDesc } from '@folio/stripes-erm-components';
 export { filterObjectKeys, filterIgnoreObjectKeys } from './filterObjectKeys';
