+++ src/hooks/index.js
@@ -1,7 +1,7 @@
 export { default as useAcqMethods } from './useAcqMethods';
 export { default as useAgreementsContexts } from './useAgreementsContexts';
 export { default as useAgreementsHelperApp } from './useAgreementsHelperApp';
-export { default as useAgreementsSettings } from './useAgreementsSettings';
+export { default as useAgreementsDisplaySettings } from './useAgreementsDisplaySettings';
 export { default as useChunkedOrderLines } from './useChunkedOrderLines';
 export { default as useSuppressFromDiscovery } from './useSuppressFromDiscovery';
 export { default as useAddFromBasket } from './useAddFromBasket';
