+++ src/routes/AgreementLinesRoute/AgreementLinesRoute.test.js
@@ -8,7 +8,7 @@ import mockRefdata from '../../../test/jest/refdata';
 
 jest.mock('../../hooks', () => ({
   ...jest.requireActual('../../hooks'),
-  useAgreementsSettings: jest.fn(() => ({ data: { configs: [{}] } })),
+  useAgreementsDisplaySettings: jest.fn(() => ({ data: { configs: [{}] } })),
   useAgreementsRefdata: () => mockRefdata,
 }));
